% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildCubeFromDataFrames.R
\name{BuildCubeFromDataFrames}
\alias{BuildCubeFromDataFrames}
\title{Build a RDF data cube based on specification}
\usage{
BuildCubeFromDataFrames(cubeMetadata, obsData, common.prefixes = NULL,
  endpoint = NULL)
}
\arguments{
\item{cubeMetadata}{data.frame with metadata}

\item{obsData}{data.frame with observations}

\item{common.prefixes}{data.frame commom prefixes, if NULL a built in default value is used}

\item{endpoint}{Used for determined codelist for dimensions. When
codetype="SDTM" to give the URL for the remote endpoint. If NULL
then the local rdf.cdisc.store from the environment is used.}
}
\value{
The filename for the generated turtle file
}
\description{
Build a RDF data cube based on specification
}

