% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildCubeFromWorkbook.R
\name{BuildCubeFromWorkbook}
\alias{BuildCubeFromWorkbook}
\title{Build a RDF data cube based on specification given in a spreadsheet}
\usage{
BuildCubeFromWorkbook(RDFCubeWorkbook, domainName, endpoint = NULL)
}
\arguments{
\item{RDFCubeWorkbook}{Filename and path to an workbook specifying the datacube}

\item{domainName}{The domainName create the data cube using the corresponding tab in the workbook}

\item{endpoint}{Used for determined codelist for dimensions. When
codetype="SDTM" to give the URL for the remote endpoint. If NULL
then the local rdf.cdisc.store from the environment is used.}
}
\value{
The filename for the generated turtle file
}
\description{
Build a RDF data cube based on specification given in a spreadsheet
}

