% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize string for use in rdfs definitions (not good version)}
\usage{
capitalize(s, strict = FALSE)
}
\arguments{
\item{s}{input string}

\item{strict}{if all character should be capitalized}
}
\value{
s capitalized
}
\description{
Capitalize string for use in rdfs definitions (not good version)
}

