% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.default.crnd.prefixes.R
\name{Get.default.crnd.prefixes}
\alias{Get.default.crnd.prefixes}
\title{Get list of default prefixes}
\usage{
Get.default.crnd.prefixes()
}
\value{
List with default prefixes
}
\description{
Get list of default prefixes
}

