% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.qb.crnd.prefixes.R
\name{Get.qb.crnd.prefixes}
\alias{Get.qb.crnd.prefixes}
\title{Add prefixes to a rrdf store}
\usage{
Get.qb.crnd.prefixes(domainName, namespaceStem = "http://www.example.org/dc/",
  domain.prefixes = c("dccs", "ds"), general.prefixes = c("code"))
}
\arguments{
\item{domainName}{The domainName to create prefixes for}

\item{namespaceStem}{The stem for the namespace}

\item{domain.prefixes}{Prefix specific for the domain}

\item{general.prefixes}{General prefixes not specific for the domain}
}
\value{
Data frame with column prefix and namespace
}
\description{
The function defines the domain specific prefixes used for the rdf
data cube for the Clinical Research and Development Analysis
Results Data
}

