% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDescrStatProcedure.R
\name{GetDescrStatProcedure}
\alias{GetDescrStatProcedure}
\title{Get list of Descriptive statistics}
\usage{
GetDescrStatProcedure()
}
\value{
List providing function for each descriptive statistics
}
\description{
For q1 and q3 type=2 is used, as it gives same results as SAS. However,
the R help on quantile states type=3 gives same result as SAS.
TODO(MJA): investigate which definition is the preferable.
}

