% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDimAttrMeasInWorkbookFormatSparqlQuery.R
\name{GetDimAttrMeasInWorkbookFormatSparqlQuery}
\alias{GetDimAttrMeasInWorkbookFormatSparqlQuery}
\title{SPARQL query for dimensions, attributes and measure from RDF data cube in workbook format}
\usage{
GetDimAttrMeasInWorkbookFormatSparqlQuery(forsparqlprefix)
}
\arguments{
\item{forsparqlprefix}{PREFIX part of SPARQL query}
}
\value{
SPARQL query
}
\description{
SPARQL query for dimensions, attributes and measure from RDF data cube in workbook format
}

