% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDomainNameFromCube.R
\name{GetDomainNameFromCube}
\alias{GetDomainNameFromCube}
\title{Get domain name for the RDF data cube
TODO: change the way domainname is used - should only use the dsdName}
\usage{
GetDomainNameFromCube(store, forsparqlprefixcommon = GetForSparqlPrefix())
}
\arguments{
\item{store}{RRDF store}

\item{forsparqlprefixcommon}{Common prefixes for the data cube}
}
\value{
the domainName for the cube
}
\description{
Get domain name for the RDF data cube
TODO: change the way domainname is used - should only use the dsdName
}

