% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetEmptyMetadata.R
\name{GetEmptyMetadata}
\alias{GetEmptyMetadata}
\title{Get empty data.frame with column corresponding to the metadata data.frame}
\usage{
GetEmptyMetadata()
}
\value{
empty data.frame with column corresponding to the metadata data.frame
}
\description{
Get empty data.frame with column corresponding to the metadata data.frame
}

