% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetForSparqlPrefix.as.df.R
\name{GetForSparqlPrefix.as.df}
\alias{GetForSparqlPrefix.as.df}
\title{Get the dataframe with prefix}
\usage{
GetForSparqlPrefix.as.df(domainName = NULL, common.prefixes = NULL,
  custom.prefixes = NULL)
}
\arguments{
\item{domainName}{Domain Name for the RDF data cube, none if NULL}

\item{common.prefixes}{Common prefixes, using default if NULL}

\item{custom.prefixes}{Custom prefixes, using default if NULL}
}
\value{
Character string with PREFIX statements delimited by CR
}
\description{
Get the dataframe with prefix
}

