% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetForSparqlPrefix.R
\name{GetForSparqlPrefix}
\alias{GetForSparqlPrefix}
\title{Get the PREFIX part for a SPARQL for the given domainname}
\usage{
GetForSparqlPrefix(domainName = NULL, common.prefixes = NULL,
  custom.prefixes = NULL)
}
\arguments{
\item{domainName}{Domain Name for the RDF data cube, none if NULL}

\item{common.prefixes}{Common prefixes, using default if NULL}

\item{custom.prefixes}{Custom prefixes, using default if NULL}
}
\value{
Character string with PREFIX statements delimited by CR
}
\description{
Get the PREFIX part for a SPARQL for the given domainname
}

