% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMetaDataInWorkbookFormatSparqlQuery.R
\name{GetMetaDataInWorkbookFormatSparqlQuery}
\alias{GetMetaDataInWorkbookFormatSparqlQuery}
\title{SPARQL query for metadata from RDF data cube in workbook format}
\usage{
GetMetaDataInWorkbookFormatSparqlQuery(forsparqlprefix)
}
\arguments{
\item{forsparqlprefix}{PREFIX part of SPARQL query}
}
\value{
SPARQL query
}
\description{
SPARQL query for metadata from RDF data cube in workbook format
}

