% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetObservationsSparqlQuery.R
\name{GetObservationsSparqlQuery}
\alias{GetObservationsSparqlQuery}
\title{SPARQL query for observations from RDF data cube in workbook format}
\usage{
GetObservationsSparqlQuery(forsparqlprefix, domainName, dimensions, attributes)
}
\arguments{
\item{forsparqlprefix}{PREFIX part of SPARQL query}

\item{domainName}{domainName for the RRDF cube}

\item{dimensions}{dimensions, vector or r x 1 matrix}

\item{attributes}{attributes, vector or r x 1 matrix}
}
\value{
SPARQL query
}
\description{
SPARQL query for observations from RDF data cube in workbook format
}

