% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetObservationsWithDescriptionSparqlQuery.R
\name{GetObservationsWithDescriptionSparqlQuery}
\alias{GetObservationsWithDescriptionSparqlQuery}
\title{SPARQL query for dimensions and attributes from RDF data cube in workbook format}
\usage{
GetObservationsWithDescriptionSparqlQuery(forsparqlprefix, domainName,
  dimensions, attributes)
}
\arguments{
\item{forsparqlprefix}{PREFIX part of SPARQL query}

\item{domainName}{domainName for the RRDF cube}

\item{dimensions}{dimensions}

\item{attributes}{attributes}
}
\value{
SPARQL query
}
\description{
SPARQL query for dimensions and attributes from RDF data cube in workbook format
}

