% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetValueFromMetadata.R
\name{GetValueFromMetadata}
\alias{GetValueFromMetadata}
\title{Get compLabel value from meta data data.frame using compType and compName as keys}
\usage{
GetValueFromMetadata(metaData, compType = NULL, compName = NULL,
  defaultValue = NULL)
}
\arguments{
\item{metaData}{data.frame with columns compType, compName and compLabel}

\item{compType}{a vector with values of compType, or NULL if not used}

\item{compName}{select on the value of compName}

\item{defaultValue}{if not NULL then a character string providing the default value in}
}
\value{
the value of the meta data or the the provided default value if keys are not found
}
\description{
Get compLabel value from meta data data.frame using compType and compName as keys
}

