% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qb.def.prefixlist.R
\name{qb.def.prefixlist}
\alias{qb.def.prefixlist}
\title{Add prefixes to a rrdf store and returns list of prefixes}
\usage{
qb.def.prefixlist(store = NULL, prefixes)
}
\arguments{
\item{store}{A rrdf store, if NULL then not added to a store}

\item{prefixes}{A data.frame with column prefix and namespace}
}
\value{
The list with member names prefixUPPERCASEPREFIX and namespace as value
TODO(mja): this should be changed - at least the naming, or use another data structure
}
\description{
Sideeffect: the prefixes are added to the store
}

