% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rrdfqbcrnd0.R
\docType{data}
\name{qbCDISCprefixes}
\alias{qbCDISCprefixes}
\title{Default PREFIXes for SPARQL queries}
\format{A list of PREFIXes for SPARQL queries}
\description{
Default PREFIXes for SPARQL queries
}
\keyword{datasets}

