% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/encodetouir.R
\name{encodetouri}
\alias{encodetouri}
\title{Encode string to URI}
\usage{
encodetouri(s)
}
\arguments{
\item{s}{A value to be used as URI}
}
\value{
The string{s} encoded as URI
}
\description{
This function is currently only a gsub - should be more clever and make better URI encoding.
}

