% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDescrStatProcedure.R
\name{GetDescrStatProcedure}
\alias{GetDescrStatProcedure}
\title{Get list of Descriptive statistics}
\usage{
GetDescrStatProcedure()
}
\value{
List providing function for each descriptive statistics
}
\description{
Get list of Descriptive statistics
}

